seperating js views
