+++ src/assets/scss/app.scss
@@ -81,8 +81,9 @@
 @import "templates/front";
 @import "templates/kitchen-sink";
 
-.expand {
+.front-hero .marketing.expand {
     transition: 500ms all ease;
     transform: scale(1.1);
+    opacity: 1;
 }
